Fix tz warning for timestamps in UTC (#4)
**Summary**
Fixing the use of `utcoffset()` to determine if a timestamp is timezone
naive or not.

**Problem Description**
When using the `get_timeseries()` method with timestamps in UTC, the
warning "Passing naive datetime is discouraged, assuming local
timezone." is shown even though the timestamps have timezone
information.

![Bildschirmfoto 2023-03-07 um 16 50
08](https://user-images.githubusercontent.com/33221251/223475068-de821268-a8b2-4772-be6a-9047cf174095.png)

**Solution**
This is because for timestamps in UTC `not time.utcoffset()` evaluates
to True. We must check if it's `None` instead.