Implement `ApiBaseClient` and derive `TimeseriesApiClient` from it (#14)
:point_up: This PR is the first in a row in an effort to expose
production runs via the ENLYZE SDK.

At the moment, we only expose time series data via the SDK and
interaction with the timeseries service is contained in the
`TimeseriesApiClient`. However, the new production runs API introduces a
different pagination pattern, which makes it inconvenient to generalize
the `TimeseriesApiClient` to a `EnlyzePlatformApiClient` :tm:.
Thus, this and subsequent PRs are building on top of the assumption that
we will have separate API Clients for the timeseries and production runs
API.

This first PR introduces an `ApiBaseClient` class, which implement
generalized `get` and `get_paginated` methods. The latter calls two
abstract methods, `_has_more` and `_next_page_call_args` that must be
implemented by API clients deriving from this class.

---------

Co-authored-by: Daniel Krebs <daniel-k@users.noreply.github.com>