Exclude always-NA columns in ProductionRuns DataFrame.
The `ProductionRun` dataclass contains some fields that in turn are dataclasses
themselves. When such fields are turned into DataFrame columns, they are
flattened such that every field contained in the sub-objects becomes a column
in the final DataFrame. Since some of these fields are actually optional, also
the field name itself will turn into a column which in practice always is
`None`. This PR removes such useless columns from the DataFrame.