Fix PR-comment GitHub Action by using our forked version. (#36)
I had to remove a problematic type annotation in the docstring of
`EnlyzeClient.get_appliances()` as well, because that has led to a
failure building the docs. For consistency, I've also removed a similar
annotation in the same class. Both annotation are not strictly
necessary, so this shouldn't be a big deal.