Enforce schema of ProductionRuns DataFrame. (#35)
The `ProductionRun` dataclass contains some fields that in turn are
dataclasses themselves. When such fields are turned into DataFrame
columns, they are flattened such that every field contained in the
sub-objects becomes a column in the final DataFrame. Since some of these
fields are actually optional, also the field name itself will turn into
a column which in practice always is `None`. This PR removes such
useless columns from the DataFrame.

Moreover, if any of the optional dataclass fields is empty for all
production runs, then their sub fields would not become part of the
DataFrame. This can be annoying when building services on top of the
SDK, because you cannot assume that a column would be part of the
DataFrame, thus requiring checks everywhere. Therefore, we now enforce
the schema of the returned DataFrame. This means, that irrespective of
the underlying data, all nested fields will be always be present as
flattened columns in the DataFrame.