Fix an issue where the SDK fails with a `TimeseriesServiceError` if querying time series data for more than 100 variables (#40)
closes https://github.com/enlyze/enlyze-python/issues/7

This PR fixes an issue where the user was allowed to request timeseries
data for any number of variables but it'd fail because the `timeseries`
API accepts a maximum of `100` variables per request. We now chunk the
variables requested into chunks of size `<=100` and issue a request for
each chunk and then merge the data returned in all of the responses.

This PR also refactors the logic to fetch timeseries data so that it has
less repetition. Previously, there was a lot of overlap in the code in
the `get_timeseries` and `get_timeseries_with_resampling` methods.