Fix incompatibility with `httpx>=0.28` (#52)
In `httpx==0.28.0`, the handling of GET query parameters has changed in
a backwards-incompatible way. The new behaviour is to replace query
parameters in the URL when parameters are passed via the `params` kwarg.
Our pagination code expects that parameters in the URL get merged with
the additional parameters of the `params` kwarg. This PR explicitly
implements the old behaviour of merging params to fix compatibility with
recent releases of `httpx`.

refs: 
- https://github.com/encode/httpx/pull/3364
- https://github.com/encode/httpx/pull/3440